add form with validation functionality
This commit creates a module called signup.js which exports an object
called Signup. This Signup object contains properties and methods
used in form validation. Whenever the user inputs a key into an input
field the form will validate itself and let the user know if there any
corrections that need to be made.